src/third_party/blink/web_tests/external/wpt/css/css-contain/parsing/contain-computed-expected.txt

This is a testharness.js-based test.
PASS Property contain value 'none' computes to 'none'
PASS Property contain value 'strict' computes to 'strict'
PASS Property contain value 'content' computes to 'content'
PASS Property contain value 'size' computes to 'size'
PASS Property contain value 'layout' computes to 'layout'
PASS Property contain value 'style' computes to 'style'
PASS Property contain value 'paint' computes to 'paint'
PASS Property contain value 'size layout' computes to 'size layout'
PASS Property contain value 'style paint' computes to 'style paint'
FAIL Property contain value 'layout style paint' computes to 'layout style paint' assert_equals: expected "layout style paint" but got "content"
FAIL Property contain value 'size layout style paint' computes to 'size layout style paint' assert_equals: expected "size layout style paint" but got "strict"
Harness: the test ran to completion.
